Battery gauge: establish that M5Dial cannot report battery state
Measured on hardware rather than assumed, in three steps:

- M5.Power.getBatteryLevel() returns -2 and getBatteryVoltage() returns
  0. M5Unified configures no battery sensing for board_M5Dial; its
  pin-identical sibling the M5DinMeter is configured, reading a 1:2
  divider on GPIO10.
- GPIO10 is the only ADC1 pin M5Dial leaves free (GPIO4-9 drive the
  GC9A01 display, GPIO1-3 are taken), so it was the one candidate worth
  probing. It reads 4095 counts -- saturated at the 3.3 V rail. A 4.1 V
  cell behind a 1:2 divider would read ~2050 mV / ~2540 counts, so it is
  not a battery sense line.
- ADC2 is unusable because WiFi owns it.

So the gauge is kept exactly as written and stays hidden on this board,
which is its designed behaviour for an implausible reading. It will
appear unchanged if a source ever reports through M5.Power -- a future
library revision, or an I2C fuel-gauge unit on Port A. The GPIO10 read
added while testing that hypothesis is removed rather than left in
reading an unrelated pin.

Also documents that USB serial only enumerates at boot: with a battery
fitted the device boots on battery power, so plugging USB in afterwards
does not bring the port up and the board needs a reset with the cable
already attached. That cost a diagnostic round trip here.

+++ src/radar.cpp
@@ -694,7 +694,7 @@ void RadarDisplay::pollBattery() {
     _battReadMs = now;
 
     int32_t level = M5.Power.getBatteryLevel();
-    int16_t mv    = M5.Power.getBatteryVoltage();
+    int32_t mv    = M5.Power.getBatteryVoltage();
 
     // A single LiPo cell: ~3.3 V effectively empty, ~4.2 V full. Anything
     // outside this band is not a cell we're reading.
